requirements.txt
astroid==2.1.0
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-cors-headers==2.4.0
django-filter==2.1.0
django-optimized-image==0.1.0
django-phonenumber-field==2.1.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
drfapikey==0.0.3
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.3.12
oauthlib==3.0.0
phonenumbers==8.10.3
Pillow==5.4.1
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
tinify==1.5.1
typed-ast==1.2.0
urllib3==1.24.1
wrapt==1.11.1
